latest microbundle, fix types
