+++ package.json
@@ -2,7 +2,8 @@
   "name": "next-themes",
   "version": "0.1.2-beta.0",
   "main": "./dist/index.js",
-  "module": "./dist/index.modern.js",
+  "module": "./dist/index.module.js",
+  "exports": "./dist/index.modern.mjs",
   "types": "./dist/index.d.ts",
   "source": "./src/index.tsx",
   "license": "MIT",
@@ -33,7 +34,7 @@
     "@types/react": "^16.9.53",
     "babel-jest": "^26.6.3",
     "jest": "^26.6.3",
-    "microbundle": "^0.12.3",
+    "microbundle": "^0.15.0",
     "prettier": "^2.2.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
